refac: refactor authentication flow
- Moved SecurityConfig and AuthController to their respective packages
- Created CustomUserDetails and CustomUserDetailsService for better user
 management
- Updated login and registration endpoints to use new user details
implementation
- Modified templates to reflect new endpoint paths
- Removed redundant code from UserServiceImpl
