+++ src/main/resources/templates/login.html
@@ -30,7 +30,7 @@
             </div>
         </form>
         <br>
-        <p>Don't have an account? <a th:href="@{/auth/register}">Register here</a></p>
+        <p>Don't have an account? <a th:href="@{/user/register}">Register here</a></p>
     </div>
 </div>
 </body>
