+++ src/test/java/ru/vladshi/cloudfilestorage/service/UserServiceImplTest.java
@@ -9,7 +9,6 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.ContextConfiguration;
@@ -109,18 +108,18 @@ public class UserServiceImplTest {
         assertThrows(UserRegistrationException.class, () -> userService.registerUser(userWithDuplicateUsername));
     }
 
-    @Test
-    @DisplayName("Should load user by username successfully")
-    public void shouldLoadUserByUsernameSuccessfully() {
-
-        userService.registerUser(testuser);
-
-        UserDetails userDetails = userService.loadUserByUsername(TEST_USERNAME);
-
-
-        assertThat(userDetails).isNotNull();
-        assertThat(userDetails.getUsername()).isEqualTo(TEST_USERNAME);
-        assertThat(userDetails.getPassword()).isEqualTo(testuser.getPassword());
-        assertThat(userDetails.getAuthorities()).isEmpty();
-    }
+//    @Test // TODO убрать
+//    @DisplayName("Should load user by username successfully")
+//    public void shouldLoadUserByUsernameSuccessfully() {
+//
+//        userService.registerUser(testuser);
+//
+//        UserDetails userDetails = userService.loadUserByUsername(TEST_USERNAME);
+//
+//
+//        assertThat(userDetails).isNotNull();
+//        assertThat(userDetails.getUsername()).isEqualTo(TEST_USERNAME);
+//        assertThat(userDetails.getPassword()).isEqualTo(testuser.getPassword());
+//        assertThat(userDetails.getAuthorities()).isEmpty();
+//    }
 }
